% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamCov.R
\name{bedtoolsgenomecov2bamCov}
\alias{bedtoolsgenomecov2bamCov}
\title{bedtoolsgenomecov2bamCov}
\usage{
bedtoolsgenomecov2bamCov(gencovdir = NULL, lvls = c(1, 5, 10, 25, 50,
  75, 100), bpwindow = 10000)
}
\description{
bedtoolsgenomecov2bamCov
}
