% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDprimer_design.R
\name{MIDPrimerFinder}
\alias{MIDPrimerFinder}
\title{MID Primer Design}
\usage{
MIDPrimerFinder(design_fasta = NULL, MID2MIDmatchesAllowed = NULL,
  MIDlength = NULL, MIDhomopolymerallowance = NULL,
  MID2targetmismatchesAllowed = NULL, MIDnum = NULL)
}
\arguments{
\item{design_fasta}{is expected to have the forward, reverse, and target gene in it. Must have `forward` and `reverse` in the contig names for the forward and reverse primers}

\item{MID2targetmismatchesAllowed}{is the number of base-pair mismatches allowed in the MID before it would bind to ANY place on the target gene}
}
\description{
This will produces MIDs
}
\details{
This is a brute-force approach with recursion as the number of new MIDs proposed is based on "how close we are to solving the problem" with the current sit of MIDs
}
\author{
Nick Brazeau
}
