% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSeekDeepfunctions.R
\name{SeekDeepDat2ExonAnnotation}
\alias{SeekDeepDat2ExonAnnotation}
\title{SeekDeepDat2ExonAnnotation}
\usage{
SeekDeepDat2ExonAnnotation(input, gff, geneid, ampliconrefseqpath,
  forwardprimerpath, reverseprimerpath, ncbigeneticcode = 1)
}
\arguments{
\item{skdpclustinfo_df_simp}{from skdp_filter_simplifier}
}
\description{
annotates exons
}
