% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSeekDeepfunctions.R
\name{SeekDeepDat2HapPlotter}
\alias{SeekDeepDat2HapPlotter}
\title{SeekDeepDat2HapPlotter}
\usage{
SeekDeepDat2HapPlotter(input, target = "Target")
}
\arguments{
\item{input}{from skdp_filter_simplifier}
}
\description{
plots
}
