% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSeekDeepfunctions.R
\name{SeekDeepOutput2SeekDeepDat}
\alias{SeekDeepOutput2SeekDeepDat}
\title{skdp_filter_simplifier}
\usage{
SeekDeepOutput2SeekDeepDat(skdpclustinfo_df, readcountcutoff = 0)
}
\arguments{
\item{skdpclustinfo_df}{from seekdeep}
}
\description{
filters
}
