% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_vcf_filters.R
\name{vcffilter_CHROMPOS}
\alias{vcffilter_CHROMPOS}
\title{vcfR2vcffilter_CHROMPOS}
\usage{
vcffilter_CHROMPOS(vcfRobject = NULL, chromposdf = NULL)
}
\arguments{
\item{chromposdf}{an dataframe that contains loci to be excluded and has been produced by \code{GFF2VariantAnnotation_Short}}
}
\description{
Produces a subsetted \code{vcfR} with specific loci excluded as determined \code{chromposdf}.
}
